+++ utils/error.ts
@@ -1,6 +1,6 @@
 // Excel file validation errors
 export const EXCEL_VALIDATION_ERROR =
-  "Tarkista että ELSA:sta hakemasi excel-tiedoston sarakkeita ei ole muokattu ja että tarvittavat sarakkeet on tiedostossa (Sarja, Pvm, Klo, Kenttä, Koti, Vieras)."
+  "Tarkista että eLSA:sta hakemasi excel-tiedoston sarakkeita ei ole muokattu ja että tarvittavat sarakkeet on tiedostossa (Sarja, Pvm, Klo, Kenttä, Koti, Vieras)."
 export const EXCEL_DATE_FORMAT_ERROR = (date: string | number) =>
   `Odottamaton päivämäärämuoto: ${date}`
 export const EXCEL_FILE_MISSING_ERROR = "Ei lisättyä tiedostoa"